labfiles/foundry-chat/python/chat-app/requirements.txt
python-dotenv
aiohttp
azure-identity
openai